Trace pylint output with the system tag
Prompt: "Title: Trace pylint output with the system tag
Description: None"
